Move release_dump into models.py and update JSONs after API POST
